[add] modular monolith.
