+++ go/cmd/db/seed/main.go
@@ -1,8 +1,8 @@
 package main
 
 import (
-	"github.com/ispec-inc/monorepo/go/pkg/infra/entity"
-	"github.com/ispec-inc/monorepo/go/pkg/mysql"
+	"github.com/yamad07/monorepo/go/pkg/infra/entity"
+	"github.com/yamad07/monorepo/go/pkg/mysql"
 	"gorm.io/gorm"
 )
