+++ go/go.mod
@@ -1,4 +1,4 @@
-module github.com/ispec-inc/monorepo/go
+module github.com/yamad07/monorepo/go
 
 go 1.16
 
@@ -9,11 +9,12 @@ require (
 	github.com/getsentry/sentry-go v0.10.0
 	github.com/go-chi/chi v1.5.4
 	github.com/golang/mock v1.5.0
+	github.com/ispec-inc/monorepo/go v0.0.0-20210503000641-7dbc94e2619f
 	github.com/k0kubun/pp v2.4.0+incompatible
 	github.com/lib/pq v1.10.0 // indirect
 	github.com/pkg/errors v0.9.1
 	github.com/stretchr/testify v1.7.0
-	go.uber.org/multierr v1.6.0 // indirect
+	go.uber.org/multierr v1.6.0
 	google.golang.org/protobuf v1.26.0
 	gorm.io/driver/mysql v1.0.5
 	gorm.io/gorm v1.21.8
