+++ go/svc/admin/cmd/server/rest/router.go
@@ -1,15 +1,25 @@
-package main
+package rest
 
 import (
 	"net/http"
 
 	"github.com/go-chi/chi"
 
-	"github.com/ispec-inc/monorepo/go/pkg/presenter"
-	v1 "github.com/ispec-inc/monorepo/go/svc/admin/pkg/controller/rest/v1"
+	"github.com/yamad07/monorepo/go/pkg/presenter"
+	v1 "github.com/yamad07/monorepo/go/svc/admin/pkg/controller/rest/v1"
+	"github.com/yamad07/monorepo/go/svc/admin/pkg/database"
+	"github.com/yamad07/monorepo/go/svc/admin/pkg/logger"
 )
 
-func NewRouter() http.Handler {
+func NewRouter() (http.Handler, func() error, error) {
+	if err := database.Init(nil); err != nil {
+		return nil, nil, err
+	}
+	cleanup, err := logger.Init()
+	if err != nil {
+		return nil, nil, err
+	}
+
 	r := chi.NewRouter()
 	r = commonMiddleware(r)
 
@@ -19,5 +29,5 @@ func NewRouter() http.Handler {
 		presenter.Response(w, map[string]string{"messsage": "ok"})
 	})
 
-	return r
+	return r, cleanup, nil
 }
