+++ go/svc/admin/pkg/controller/rest/v1/article/controller_test.go
@@ -5,11 +5,11 @@ import (
 	"net/http/httptest"
 	"testing"
 
-	"github.com/ispec-inc/monorepo/go/pkg/infra/entity"
-	"github.com/ispec-inc/monorepo/go/pkg/testool"
-	pb "github.com/ispec-inc/monorepo/go/proto/admin/api/rest/v1/article"
-	"github.com/ispec-inc/monorepo/go/svc/admin/pkg/database"
 	"github.com/stretchr/testify/assert"
+	"github.com/yamad07/monorepo/go/pkg/infra/entity"
+	"github.com/yamad07/monorepo/go/pkg/testool"
+	pb "github.com/yamad07/monorepo/go/proto/admin/api/rest/v1/article"
+	"github.com/yamad07/monorepo/go/svc/admin/pkg/database"
 )
 
 func TestArticleController_list(t *testing.T) {
