+++ go/svc/admin/pkg/database/database.go
@@ -1,9 +1,11 @@
 package database
 
 import (
+	"log"
+
 	"gorm.io/gorm"
 
-	"github.com/ispec-inc/monorepo/go/pkg/mysql"
+	"github.com/yamad07/monorepo/go/pkg/mysql"
 )
 
 var db *gorm.DB
@@ -18,5 +20,6 @@ func Init(indb *gorm.DB) (err error) {
 }
 
 func Get() *gorm.DB {
+	log.Println(db)
 	return db
 }
