+++ go/svc/article/cmd/server/rest/router.go
@@ -1,16 +1,33 @@
-package main
+package rest
 
 import (
 	"net/http"
 
 	"github.com/go-chi/chi"
 
-	"github.com/ispec-inc/monorepo/go/pkg/presenter"
-	v1 "github.com/ispec-inc/monorepo/go/svc/article/pkg/adapter/rest/v1"
-	"github.com/ispec-inc/monorepo/go/svc/article/pkg/registry"
+	"github.com/yamad07/monorepo/go/pkg/presenter"
+	v1 "github.com/yamad07/monorepo/go/svc/article/pkg/adapter/rest/v1"
+	"github.com/yamad07/monorepo/go/svc/article/pkg/registry"
 )
 
-func NewRouter(rgst registry.Registry) http.Handler {
+func NewRouter() (http.Handler, func() error, error) {
+	repo, repoCleanup, err := registry.NewRepository()
+	if err != nil {
+		return nil, nil, err
+	}
+
+	lgr, lgrCleanup, err := registry.NewLogger()
+	if err != nil {
+		return nil, nil, err
+	}
+	cleanup := func() error {
+		repoCleanup()
+		lgrCleanup()
+		return nil
+	}
+
+	rgst := registry.NewRegistry(repo, lgr)
+
 	r := chi.NewRouter()
 
 	r = commonMiddleware(r)
@@ -21,5 +38,5 @@ func NewRouter(rgst registry.Registry) http.Handler {
 		presenter.Response(w, map[string]string{"messsage": "ok"})
 	})
 
-	return r
+	return r, cleanup, err
 }
