+++ go/svc/article/pkg/domain/repository/user.go
@@ -1,7 +1,7 @@
 package repository
 
 import (
-	"github.com/ispec-inc/monorepo/go/svc/article/pkg/domain/model"
+	"github.com/yamad07/monorepo/go/svc/article/pkg/domain/model"
 )
 
 type User interface {
