+++ go/svc/article/pkg/infra/dao/user.go
@@ -1,9 +1,9 @@
 package dao
 
 import (
-	"github.com/ispec-inc/monorepo/go/pkg/apperror"
-	"github.com/ispec-inc/monorepo/go/pkg/infra/entity"
-	"github.com/ispec-inc/monorepo/go/svc/article/pkg/domain/model"
+	"github.com/yamad07/monorepo/go/pkg/apperror"
+	"github.com/yamad07/monorepo/go/pkg/infra/entity"
+	"github.com/yamad07/monorepo/go/svc/article/pkg/domain/model"
 	"gorm.io/gorm"
 )
