+++ go/svc/article/pkg/registry/repository.go
@@ -3,9 +3,9 @@ package registry
 import (
 	"gorm.io/gorm"
 
-	"github.com/ispec-inc/monorepo/go/pkg/mysql"
-	"github.com/ispec-inc/monorepo/go/svc/article/pkg/domain/repository"
-	"github.com/ispec-inc/monorepo/go/svc/article/pkg/infra/dao"
+	"github.com/yamad07/monorepo/go/pkg/mysql"
+	"github.com/yamad07/monorepo/go/svc/article/pkg/domain/repository"
+	"github.com/yamad07/monorepo/go/svc/article/pkg/infra/dao"
 )
 
 type Repository struct {
